add a new folder named Components, add import and export welcome function to main
